Huge update, warps can now have same name

+++ src/net/plommer/UltraWarp/Commands/RemoveCommand.java
@@ -1,5 +1,7 @@
 package net.plommer.UltraWarp.Commands;
 
+import java.util.ArrayList;
+
 import net.plommer.UltraWarp.Utils;
 import net.plommer.UltraWarp.Warps;
 
@@ -15,19 +17,23 @@ public class RemoveCommand extends BaseCommand {
 	
 	@Override
 	public boolean execute() {
-		Warps w = plugin.db.checkWarp(args.get(0));
-		if(w != null) {
-			if(w.getPlayerUUID().equals(player.getUniqueId()) || player.hasPermission(permission() + "admin")) {
-				if(plugin.db.removeWarps(w)) {
-					Utils.sendMessage(sender, "&a" + w.getWarpName() + " &eHas been deleted!");
+		ArrayList<Warps> warp = plugin.db.checkWarp(args.get(0));
+		if(warp.size() > 0) {
+			for(Warps w : warp) {
+				if(w != null) {
+					if(w.getPlayerUUID().equals(player.getUniqueId()) || player.hasPermission(permission() + "admin")) {
+						if(plugin.db.removeWarps(w)) {
+							Utils.sendMessage(sender, "&a" + w.getWarpName() + " &eHas been deleted!");
+						} else {
+							Utils.sendMessage(sender, "&cUnexpected error!");
+						}
+					} else {
+						Utils.sendMessage(sender, "&cYou do not own that warp!");
+					}
 				} else {
-					Utils.sendMessage(sender, "&cUnexpected error!");
+					Utils.sendMessage(sender, "&cThat warp dosn't exist!");
 				}
-			} else {
-				Utils.sendMessage(sender, "&cYou do not own that warp!");
 			}
-		} else {
-			Utils.sendMessage(sender, "&cThat warp dosn't exist!");
 		}
 		return true;
 	}
